Fix suspending toggle

+++ kill.sh
@@ -1,9 +1,13 @@
 #!/usr/bin/env bash
 
-
-(killall light-locker) &
-(killall xidlehook) &
-notify-send 'Suspending is disabled!'
-
+if  pgrep -f xidlehook > 0 && pgrep -f light-locker > 0
+then
+    (killall light-locker) &
+    (killall xidlehook) &
+    notify-send 'Suspending is disabled!'
+else
+    notify-send 'Suspending is enabled!'
+    exec /usr/bin/sus.sh
+fi
